Split utils code into separate files (one function per file), update description of "getUniquePaths", correct typos
