+++ src/utils/index.ts
@@ -1,2 +1,3 @@
-export * from './file-path'
+export * from './filter-paths'
+export * from './get-unique-paths'
 export * from './types'
